Fix to modification of 'nonEffInterval' argument
The modification to nonEffInterval, in the case that times are provided
rather than infection counts, did not ensure that the specified times
remaining within the bounds of the trial duration.  I added code to
truncate the nonEffTimes vector that is derived so that all times are
within the run-time of the trial.

I also replaced the original lines of code with equivalent ones that
are cleaner.