Deprecated argument 'estimand' in monitorTrial
Argument 'estimand' was deprecated and is being replaced by two new
arguments, 'nonEffEst' and 'effEst', which specify the estimands to
use for non-efficacy and efficacy analyses respectively.

This is being done because non-efficacy monitoring can be done based
on both cuminc and cox estimands if 'combined' is chosen, but efficacy
analyses require a SINGLE estimand.  The code has always, silently,
used 'cuminc' for efficacy analyses (highEff and final stage 1) when
"combined" was specified, so this change allows for more clarity
around what is happening.

The documentation has NOT been updated to reflect the deprecation or
the new arguments.  Here's a short synopsis of the new arguments:

nonEffEst  Specifies the estimand(s) to use for non-efficacy monitoring
           Valid values are: 'cuminc', 'cox', 'combined'
           'combined' indicates to do monitoring using BOTH the cox and
           cumulative incidence estimands, and declare nonEff only if
           both satisfy the stopping criteria

effEst     Specifies the estimand to be used for all efficacy analyses:
           highEff (if requested), interim-efficacy (if requested), and
           the stage 1 analysis.