applyStopRules was not censoring to lagTime
When I modified applyStopRules to change how it was operating -
removing the laggedMonitoring, and do both lagged and unlagged
within a single call, I failed to make sure that censoring was
still being done in the code that remained.  It wasn't.  But
I added back code to do it, so now we're good.