Fixed issue with 'N1' in monitorTrial
The code setting N1 is very complicated - and really needs to be scrapped.
In trying to adapt it I put in some code that defined N1 conditionally,
at one point, based on whether or not it already had been set at a
certain point in the code.  That works fine the first time through the
code, but for later passes through, N1 exists from the previous iteration
and so the conditional code did not get run after the first trial, so N1
did not get reset for other trials.