Avoid warning msg in rm("N1") by checking existance
The code I introduced to remove N1 for each new trial, when a
constant N1 is not being used, threw a warning for the first trial
because N1 did not yet exist in that case.  So I changed the code
to check for existance of N1 before trying to remove it.