Fixed bug when not all test times reached
If e.g. 3 tests were requested, along with three different alpha levels,
but there weren't enough infections accrued in the trial so the 3rd test
was not done, then the imbalance between tests (2) and alphaLevels values
(3) caused each test to be run at every alpha level.

So code was added to catch this condition and to subset alphaLevel to
be only length 2 also, and prevent errors.