Modified monitorTrial behavior when multiple tests aren't reached
In cases in which not enough infections accrue during the trial to
reach the specified "test times" (specified counts at which tests
will be performed). Until recently only the tests reached would be
performed.  Then a few days ago, I modified the code to insert a
final test at the end of stage 1, if the final test was not reached,
but the code would stop if *more* than one test was not reached.
This caused problems, so the current modification changes that final
behavior, when multiple tests are not reached.

As before, a final test is inserted at the end of stage 1 follow-up,
but since we don't have alpha-spending implemented yet, I can't
determine the correct alpha level for that final test.  So the code
issues a warning and performs that test using the nominal alpha level
previously specified for the final test.   That's the best we can do
until alpha spending is implemented (and even after that, if the user
doesn't use the alpha spending mechanism).