Fixed issue with harm monitoring - wasn't correct
The modifications made to the specification and timing of
non-efficacy monitoring needed to be moved up higher in the code
so they could be used to correctly set the range for harm monitoring.

This was needed to correct incorrect code added in the last few days
that stopped harm monitoring at N1.ne rather than N1 (i.e. it stopped
harm monitoring at the count at which non-eff started, but this was
incorrect because the counts are relative to different cohorts.
That was not the intention of the code, as written, but it's how it
functioned.  This issue was fixed but, as noted, involved moving
up a large chunk of code used to process the arguments and deterimine
the timing of the non-efficacy analyses.