Replaced function pNS in monitoring_utils.R
The version of pNS that was included in seqDesign did not have a
feature that is needed, which is when the first stopping bound
is not of the form N:0, but rather  e.g. (N-4):4, then it did not
sum up all probs of attaining (N-4):4, but just the prob. of
reaching it exactly on the Nth test.

Anyway, we had a version of the code that was written for use with AMP
and that was modified to deal with this, but for some reason it didn't
make it into the package.  There are quite a few diff.s in the various
harm monitoring functions, so it's not just one function that's different.
Anyway, it was an easy fix to swap in the other version and make a small
change to the calling of the function (argument 'N' no longer needed).