Minor mods to getHarmBounds and getAlphaPerTest
After swapping out the old pNS with the new one there was an issue
with the uniroot call when it evaluated the inner function at per.test=0.
So I changed the search interval to start at 0.000001 instead and all is
good.

Changes to getHarmBound were mostly cosmetic - I commented out several
lines of code that weren't doing anything that was needed anymore.  Though
I did correct the "RR" column of the harmBounds file so that it's correct
when the randomization is other than 1:1.