Harm code fix and modification
Modification:  changed the convergence tolerance in the uniroot call
used to find the 'alphaPerTest' value to be more stringent (1e-6).
The default was not doing as good of a job as it could/should.

Fix: when I modified the harm code earlier in the week, it inadvertantly
switched the ordering of two of the columns.  But later in the code,
was this renaming:
  names(harmBounds)[1:3] <- c("N", "V", "P")

But the third column was no longer "P", it contained "RR", so "RR" got renamed
to "P" and 'Nplac' (which should have been renamed to "P"), stayed as "Nplac".
Thankfully, the code only uses columns "N" and "V", which were not messed up,
so this had no effect on the results.

I have rewritten the code so that renaming is done properly, specifying that
column 'n' is renamed to "N", column 'Nvacc' to 'V' and 'Nplac' to 'P' - with
no mention of the position of the columns.  Then I added a reordering step