non-eff monitoring updated to allow vector alpha
The non-efficacy testing in monitorTrial was changed to allow for use of
different nominal alpha values at different tests so that things like
O'Brien-Fleming type bounds can be used to control the type-I error.

Also changed during this process was the interface for specifying the
details of the non-efficacy monitoring.  Five previous arguments have
been rendered obsolete and replaced by new components within the
'nonEffCohorts' argument list.

The obsolete arguments, and their replacements, are:
  nonEffTimes    - replaced by nonEffCohorts$times
  nonEffTimeUnit - replaced by nonEffCohorts$timeUnit
  nonEffTimingCohort - replaced by nonEffCohorts$timingCohort
  upperVEnoneff  - replaced by nonEffCohorts$cohort<i>$nullVE
  alphaNoneff    - replaced by nonEffCohorts$cohort<i>$nominalAlphas

  where $cohort<i>  would be e.g. cohort1, cohort2, ...
  so these arguments are now cohort-specific, in case you wish to
  use multiple non-eff cohorts.

An example specification is given below:
---------------------------------------
 nonEffCohorts = list(
     times=NULL, timeUnit="counts",
     timingCohort=list(lagTime=2, cohortInd=NULL),

     cohort1 = list(
         estimand="cox", lagTime=6, cohortInd="pp1",
         nullVE=NULL, nominalAlphas=c(0.0030, 0.0183, 0.0440)
     )
 ),