Added new argument 'stage1Eff' to monitorTrial
Argument 'stage1Eff' will be used to specify an efficacy analysis to be
performed at the end of stage 1 follow-up.  Code implementing this
analysis has not yet been added, only a few lines of code using it so
far are present.

A second change also included (mistakenly - I meant it to be a separate
commit), is the removal of code related to options "?" and "old" for the
argument 'nonEffStartMethod' in monitorTrials.

I plan to later get rid of argument 'nonEffStartMethod' entirely, by relocating
it inside of 'nonEffCohorts', but for the moment I've only removed code that
will not be kept and relocated.